fix(topol): fix reading level description
On Windows opening a file in rt mode strips the \r character macOS does not do
this and the \r character will still be present so we need to replace any
possible \r characters.

Without this, the game will crash with buildtype=debug, because it hits a
missing codepoint in ABC8.