refactor(valaszt2): max visible entries depends on screen height
Instead of manually having to write the max number of visible entries for every
menu, automatically calculate the max number based on the current menu's
characteristics.

The only place in the menu where the number would potentially need to be
limited would be in afterplay() where the text could overwrite "You Failed to
Finish!" if we added more menu items, but at the moment it has no impact.