feat(dialog): standardize and improve editor dialog mouse behaviour
Consolidate a lot of repeated, and slightly inconsistent code into
update_and_draw_cursor().

Lots of the code was using the following condition to determine if it needed to
redraw the cursor:

    if (ujx != Moux && ujy != Mouy)

However this only redraws the cursor if the x AND y have changed. Use the
correct logic in update_and_draw_cursor(), and redraw if x OR y have changed.